fix(audit): pass 2 — keep drag_snap.cpp under 800 lines
The bug-fix carve-out pushed drag_snap.cpp from 797 to 811 lines, over the
CLAUDE.md 800-line limit. Extract callEndDrag (the self-contained drag-end
DragOutcome dispatch, ~233 lines, single caller in lifecycle.cpp) into a new
sibling TU drag_end.cpp. Pure move, no logic change; drag_snap.cpp drops to
567 lines. Build + all 282 tests green.
