+++ kwin-effect/CMakeLists.txt
@@ -56,6 +56,7 @@ set(plasmazones_effect_SRCS
     plasmazoneseffect/daemon_bringup.cpp
     plasmazoneseffect/daemon_apply.cpp
     plasmazoneseffect/drag_snap.cpp
+    plasmazoneseffect/drag_end.cpp
     plasmazoneseffect/paint_pipeline.cpp
     plasmazoneseffect/shader_transitions.cpp
     plasmazoneseffect/shader_resolve.cpp
